Move and tidy some test files around
