+++ test/parser/CMakeLists.txt
@@ -1,0 +1,3 @@
+loxmocha_add_test(parser parser-expr-test.cpp)
+
+target_link_libraries(loxmocha_parser_test PRIVATE loxmocha_parser)
